+++ packages/ui/vite.config.ts
@@ -1,7 +1,7 @@
+import tailwindcss from '@tailwindcss/vite';
 import react from '@vitejs/plugin-react';
 import { defineConfig } from 'vite';
-import { breakpointCustomMediaPlugin } from '../../scripts/breakpoint-css.js';
 
 export default defineConfig({
-  plugins: [breakpointCustomMediaPlugin(), react()],
+  plugins: [react(), tailwindcss()],
 });